Reworked various functions to include the mainMenu header for better menu pathing.

+++ project1/globalVariables.h
@@ -11,7 +11,7 @@ extern const short studentInfoSize = 6;
 extern string studentInfo[studentInfoSize];									//stores student biographical info. 6 items tied to index location.
 extern const short assignmentTotalSize = 10;								//this is a dummy value for the total number of assignments. In the next version, this should be dynamic.
 extern string assignmentNames[assignmentTotalSize] = { "undefined" };		//this array holds the assignment names.
-extern unsigned assignmentPointsPossible[assignmentTotalSize];				//this array holds the assignment values.
+extern float assignmentPointsPossible[assignmentTotalSize];				//this array holds the assignment values.
 extern float assignmentPointsScored[assignmentTotalSize];					//this array holds the student scores.
 extern char letterGrade;													//this char is for the student's overall letter grade. In the next version this will be an object within the student class.
 extern string teacherFullName;												//this string is teacher's last name and title.
